+++ src/components/signup/index.js
@@ -3,6 +3,7 @@ import React from "react";
 function SignUp() {
     return (
         <div>
+        signup
         </div>
     )
 }
